oppen: use tokens length when calling modulo on left and right
The was a time when using the scan_stack length was the right thing to
do since we were always upsizing all our arrays at the same time.
However, in the current version it breaks if the scan stack does not need
upsizing and the tokens and size lists need upsizing.